Electron removal/Game Details/CORS policy
Electron dependency was removed in favor of focusing on website functionality and later mobile app functionality.  Added a js function to get game details from steam but this caused a CORS error. I attempted to fix this error with a node.js proxy server but this caused steam to return, presumably, the html page for the app instead of the json data. To circumvent this, I will be implementing the request on the backend. The proxy server code is being left in as I believe this will be useful in a production version.